Fix argparse failure

+++ Unikraft/unikraft/lib/ukfile/file-nops.c
@@ -10,17 +10,15 @@
 
 
 ssize_t uk_file_nop_read(const struct uk_file *f __unused,
-			 const struct iovec *iov __unused,
-			 size_t iovcnt __unused,
-			 size_t off __unused, long flags __unused)
+			 const struct iovec *iov __unused, int iovcnt __unused,
+			 off_t off __unused, long flags __unused)
 {
 	return -ENOSYS;
 }
 
 ssize_t uk_file_nop_write(const struct uk_file *f __unused,
-			  const struct iovec *iov __unused,
-			  size_t iovcnt __unused,
-			  size_t off __unused, long flags __unused)
+			  const struct iovec *iov __unused, int iovcnt __unused,
+			  off_t off __unused, long flags __unused)
 {
 	return -ENOSYS;
 }
